# swagger/swagger.yaml
consumes:
- application/json
- application/xml
- application/gob
- application/x-gob
definitions:
  Messagetype:
    description: example (default view)
    example:
      id: 1
      message: ok
    properties:
      id:
        default: 0
        description: id
        example: 1
        format: int64
        type: integer
      message:
        default: ""
        description: message
        example: ok
        type: string
    required:
    - id
    - message
    title: 'Mediatype identifier: application/vnd.messagetype+json; view=default'
    type: object
  SubExamplePayload:
    example:
      message: ok
    properties:
      message:
        default: ""
        description: message
        example: ok
        type: string
    required:
    - message
    title: SubExamplePayload
    type: object
  error:
    description: Error response media type (default view)
    example:
      code: invalid_value
      detail: Value of ID must be an integer
      id: 3F1FKVRR
      meta:
        timestamp: 1.458609e+09
      status: "400"
    properties:
      code:
        description: an application-specific error code, expressed as a string value.
        example: invalid_value
        type: string
      detail:
        description: a human-readable explanation specific to this occurrence of the
          problem.
        example: Value of ID must be an integer
        type: string
      id:
        description: a unique identifier for this particular occurrence of the problem.
        example: 3F1FKVRR
        type: string
      meta:
        additionalProperties: true
        description: a meta object containing non-standard meta-information about
          the error.
        example:
          timestamp: 1.458609e+09
        type: object
      status:
        description: the HTTP status code applicable to this problem, expressed as
          a string value.
        example: "400"
        type: string
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
externalDocs:
  description: wiki
  url: ""
host: localhost:8080
info:
  contact:
    email: peeeei0804@gmail.com
    name: pei
    url: https://github.com/pei0804/goa-stater/issues
  description: pei0804/goa-stater
  license:
    name: MIT
  title: pei0804/goa-stater
  version: ""
paths:
  /examples/{id}:
    get:
      description: ID
      operationId: example#id
      parameters:
      - default: 0
        description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/vnd.goa.error
      - application/vnd.messagetype+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Messagetype'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      security:
      - generalAuth: []
      summary: id example
      tags:
      - example
  /examples/main:
    get:
      description: main example
      operationId: example#main
      produces:
      - application/vnd.goa.error
      - application/vnd.messagetype+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Messagetype'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      security:
      - generalAuth: []
      summary: main example
      tags:
      - example
  /examples/sub:
    post:
      description: sub example
      operationId: example#sub
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/SubExamplePayload'
      produces:
      - application/vnd.goa.error
      - application/vnd.messagetype+json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Messagetype'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/error'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/error'
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      security:
      - generalAuth: []
      summary: sub example
      tags:
      - example
  /swagger.json:
    get:
      operationId: swagger#/swagger.json
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
      schemes:
      - http
      summary: Download swagger/swagger.json
  /swaggerui/{filepath}:
    get:
      operationId: swaggerui#/swaggerui/*filepath
      parameters:
      - description: Relative file path
        in: path
        name: filepath
        required: true
        type: string
      responses:
        "200":
          description: File downloaded
          schema:
            type: file
        "404":
          description: File not found
          schema:
            $ref: '#/definitions/error'
      schemes:
      - http
      summary: Download swaggerui/
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
responses:
  NotFound:
    description: Not Found
  OK:
    description: OK
    schema:
      $ref: '#/definitions/Messagetype'
schemes:
- http
securityDefinitions:
  adminAuth:
    description: トークン(admin)
    in: header
    name: X-Authorization
    type: apiKey
  generalAuth:
    description: トークン(general)
    in: header
    name: X-Authorization
    type: apiKey
swagger: "2.0"